+++ src/Client/AbstractClient.php
@@ -24,61 +24,34 @@ use Psr\Log\LoggerInterface;
 
 /**
  * Class AbstractClient
- *
  * Exception code range: 2000-2049
- *
  *
  * @author Romain Cottard
  */
 class AbstractClient
 {
-    /** @var string BASE_URL */
     private const BASE_URL = 'https://mtgmelee.com';
 
-    /** @var ClientInterface $client */
     private ClientInterface $client;
-
-    /** @var RequestFactoryInterface $requestFactory */
     private RequestFactoryInterface $requestFactory;
-
-    /** @var UriFactoryInterface $uriFactory */
     private UriFactoryInterface $uriFactory;
-
-    /** @var StreamFactoryInterface $streamFactory */
     private StreamFactoryInterface $streamFactory;
 
-    /** @var LoggerInterface $logger */
-    private LoggerInterface $logger;
-
-    /**
-     * AbstractClient constructor.
-     *
-     * @param ClientInterface $client
-     * @param RequestFactoryInterface $requestFactory
-     * @param UriFactoryInterface $uriFactory
-     * @param StreamFactoryInterface $streamFactory
-     * @param LoggerInterface $logger
-     */
     public function __construct(
         ClientInterface $client,
         RequestFactoryInterface $requestFactory,
         UriFactoryInterface $uriFactory,
-        StreamFactoryInterface $streamFactory,
-        LoggerInterface $logger
+        StreamFactoryInterface $streamFactory
     ) {
         $this->client         = $client;
-        $this->logger         = $logger;
         $this->requestFactory = $requestFactory;
         $this->uriFactory     = $uriFactory;
         $this->streamFactory  = $streamFactory;
     }
 
     /**
-     * @param string $path
-     * @param FormatterInterface $formatter
-     * @param string $method
-     * @param array $params
-     * @return mixed
+     * @phpstan-param array<string, string|int|float|bool|array<string|int>> $params
+     * @phpstan-return array<object>|object|string|null
      * @throws MtgMeleeClientException
      */
     final protected function fetchResult(
@@ -88,7 +61,6 @@ class AbstractClient
         array $params = []
     ) {
         $response    = null;
-        $data        = null;
         $decodedData = null;
 
         try {
@@ -98,6 +70,7 @@ class AbstractClient
             $data = $response->getBody()->getContents();
 
             if (!empty($data)) {
+                /** @var string|\stdClass $decodedData */
                 $decodedData = json_decode($data, null, 512, JSON_THROW_ON_ERROR);
             }
 
@@ -113,50 +86,74 @@ class AbstractClient
             throw new MtgMeleeClientException('[CLI-2001] Unable to decode json response!', 2001, $exception);
         } catch (ClientExceptionInterface $exception) {
             throw new MtgMeleeClientException('[CLI-2000] ' . $exception->getMessage(), 2000, $exception);
-        } finally {
-            if (!empty($exception) && $exception instanceof \Exception) {
-                $this->getLogger()->notice($exception->getMessage(), [
-                    'type'      => 'component.mtgmelee.client.fetch',
-                    'exception' => $exception,
-                ]);
-            }
         }
 
         return $decodedData !== null ? $formatter->format($decodedData) : $decodedData;
     }
+    /**
+     * @phpstan-param array<string, string|int|float|bool|array<string|int>> $params
+     * @phpstan-return array<object>|object|string|null
+     * @throws MtgMeleeClientException|\JsonException
+     */
+    final protected function fetchPageResult(
+        string $path,
+        FormatterInterface $formatter,
+        string $method = 'GET',
+        array $params = []
+    ) {
+        $response = null;
+
+        try {
+            $request  = $this->getRequest($path, $method, $params, false);
+            $response = $this->client->sendRequest($request);
+
+            $data = $response->getBody()->getContents();
+
+            if ($response->getStatusCode() >= 400) {
+                throw new MtgMeleeClientException(); // @codeCoverageIgnore
+            }
+        // @codeCoverageIgnoreStart
+        } catch (MtgMeleeClientException $exception) {
+            $code    = 2005;
+            $message = 'Error ' . ($response !== null ? $response->getStatusCode() : ' XXX') . ' on call!';
+
+            throw new MtgMeleeClientException($message, $code, $exception);
+        } catch (ClientExceptionInterface $exception) {
+            throw new MtgMeleeClientException('[CLI-2006] ' . $exception->getMessage(), 2006, $exception);
+        }
+        // @codeCoverageIgnoreEnd
+
+        return $formatter->format($data);
+    }
 
     /**
-     * @param string $path
-     * @param string $method
-     * @param array $params
-     * @return RequestInterface
+     * @phpstan-param array<string, string|int|float|bool|array<string|int|float|bool>> $params
+     * @throws \JsonException
      */
-    protected function getRequest(string $path, string $method = 'GET', array $params = []): RequestInterface
-    {
+    protected function getRequest(
+        string $path,
+        string $method = 'GET',
+        array $params = [],
+        bool $isJsonApi = true
+    ): RequestInterface {
         $uri     = $this->uriFactory->createUri(self::BASE_URL . $path);
         $request = $this->requestFactory->createRequest($method, $uri);
 
         if (in_array($method, ['POST', 'PUT']) && isset($params['body'])) {
-            $body = is_string($params['body']) ? $params['body'] : json_encode($params['body']);
+            $body = is_string($params['body']) ? $params['body'] : json_encode($params['body'], JSON_THROW_ON_ERROR);
             $request = $request->withBody($this->streamFactory->createStream($body));
         }
 
         //~ Add header
-        return $request
-            ->withAddedHeader('Accept', 'application/json')
-        ;
+        if ($isJsonApi) {
+            $request = $request->withAddedHeader('Accept', 'application/json');
+        }
+
+        return$request;
     }
 
     /**
-     * @return LoggerInterface
-     */
-    final protected function getLogger(): LoggerInterface
-    {
-        return $this->logger;
-    }
-
-    /**
-     * @param mixed $data
+     * @param \stdClass|string|null $data
      * @param ResponseInterface|null $response
      * @return int
      */
@@ -174,12 +171,12 @@ class AbstractClient
     }
 
     /**
-     * @param \stdClass|string $data
-     * @param ResponseInterface $response
+     * @param \stdClass|string|null $data
+     * @param ResponseInterface|null $response
      * @param int $internalCode
      * @return string
      */
-    private function getErrorMessage($data, ResponseInterface $response, int $internalCode): string
+    private function getErrorMessage($data, ?ResponseInterface $response, int $internalCode): string
     {
         $error = !empty($data->errors) && is_array($data->errors) ? reset($data->errors) : null;
 
@@ -188,7 +185,7 @@ class AbstractClient
         //~ Override default prefix
         if (!empty($error->code)) {
             $prefix = '[API-' . $error->code . '] ';
-        } elseif ($response->getStatusCode() >= 400) {
+        } elseif ($response !== null && $response->getStatusCode() >= 400) {
             $prefix = '[HTTP-' . $response->getStatusCode() . '] ';
         }