+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2023-03
+[1.0.0]: https://github.com/MagicLegacy/mtgmelee-client/compare/0.4.3...1.0.0
+### Added
+- Added `TournamentClient::getTournament(): Tournament|null`
+- Added `TournamentClient::getRound(): Round[]`
+- Added entity `Tournament` & entity `Round`
+- Added tests
+- Added examples
+### Removed
+- Remove Internal serializer. Now use eureka/component-serializer
+- Remove logger injection in client
+### Changed
+- Fix phpstan error to lvl max
+- Run main analyze for PHP 8.1
+- Updating some phpdoc & cleaning other
+- Update tests
+- Update README
+
+
 ## [0.4.3] - 2023-03-10
 [0.4.3]: https://github.com/MagicLegacy/mtgmelee-client/compare/0.4.2...0.4.3
 ### Changed